fix(test): change vcs type
